+++ packages/shared/lib/core/layer-2/actions/index.ts
@@ -1,2 +1,3 @@
+export * from './getLayer2NativeTokens'
 export * from './loadEvmAddressForSelectedAccount'
 export * from './registerLayer2EventHandlers'
