+++ packages/shared/lib/core/layer-2/constants/index.ts
@@ -1,8 +1,9 @@
 export * from './accounts-contract.constant'
 export * from './contract-functions.constant'
 export * from './empty-buffer.constants'
-export * from './externally-owned-account.constant'
 export * from './externally-owned-account-type-id.constant'
+export * from './externally-owned-account.constant'
 export * from './gas-budget.constant'
+export * from './isc-contract.constant'
 export * from './target-contracts.constant'
 export * from './transfer-allowance.constant'
