+++ packages/shared/lib/core/layer-2/constants/isc-contract.constant.ts
@@ -1,0 +1,1 @@
+export const ISC_CONTRACT = '0x1074000000000000000000000000000000000000'
