+++ packages/shared/lib/core/layer-2/index.ts
@@ -1,3 +1,4 @@
+export * from './abis'
 export * from './actions'
 export * from './constants'
 export * from './classes'
