+++ packages/shared/lib/core/layer-2/interfaces/layer2-native-token.interface.ts
@@ -1,0 +1,7 @@
+import { IIrc30Metadata } from '@core/wallet'
+
+export interface ILayer2NativeToken {
+    id: string
+    amount: bigint
+    metadata?: IIrc30Metadata
+}
