+++ packages/shared/lib/core/layer-2/types/abi.type.ts
@@ -1,0 +1,1 @@
+export type AbiType = 'function' | 'constructor' | 'event' | 'fallback'
