+++ packages/shared/lib/core/layer-2/types/index.ts
@@ -1,1 +1,3 @@
+export * from './abi.type'
 export * from './layer2-metadata.interface'
+export * from './state-mutability.type'
