+++ packages/shared/lib/core/layer-2/types/state-mutability.type.ts
@@ -1,0 +1,1 @@
+export type StateMutabilityType = 'pure' | 'view' | 'nonpayable' | 'payable'
