+++ packages/shared/lib/core/network/utils/index.ts
@@ -3,7 +3,6 @@ export * from './buildPersistedNetworkFromNodeInfoResponse'
 export * from './checkNetworkId'
 export * from './checkNodeUrlValidity'
 export * from './doesNodeHavePlugin'
-export * from './getAssetsOfChain'
 export * from './getDefaultClientOptions'
 export * from './getDefaultPersistedNetwork'
 export * from './getNetworkIdFromNetworkName'
